adding the notional quote manager
The notional quote manager calculates for a given notional amount the notional based quote with a straightforward methodology. on a tick level basis. The calculation is async with the order book building, hence the first updates (while syncing with all exchanges when using multiple) may lead to crosses, that are ignored in the test() function. the output is a json quote, the test produces a file with the output and the paramters